Made creating a TextReader its own method. Refactored ReadConversation to take a reader.
Seemed to me that _reading_ a conversation was _different_ than creating a reader. So, rather than creating a reader and then reading in the same function, I’ve separated the responsibilities up.

This also means (theoretically speaking) that if a user wants a different implementation — StringReader, maybe — then they can have that as the type is TextReader. Issues with creating a reader need handling in the function.

Finally, I’ve made all of the split[x] arguments into variables in an attempt to make the Add() statement and the data being provided to the constructor clearer.