New GetStreamWriter. Remove ExportConversation in favour of WriteConversation. Decorate exceptions with more detail.
As with the reader, the writer was being made at the same time as writing a document. The creation of the writer is different from the writer itself and I wanted to make that clearer in the code. Also better to create just one writer rather than newing one up every time you call that function.

ExportConversation became an unnecessary middle function that wasn’t adding anything meaningful. Instead of keeping it, I decided to delete and simplify the sequence: get arguments, read file, write file. There will need to be some parsing of the validation / actual validation but that's still to come.

Exceptions: attempted to decorate with more information. If a file wasn’t found, I say which file. If a file can’t be opened, I say which file that is in a bid to make it clearer.

The rather large test remains working and is to be refactored shortly...