Extract  lines in ExtractConversation based on a set of rules. The idea here is that we pull out a line and, depending on the rules in the config, decide whether to add to the final output or not.
Right now, the checks for validation are in a bunch of if statements. I don’t like that and will be extracting all that logic into a function.

Each validation-based function takes a message, a string and returns a bool so I’m thinking:
- Refactor the validation funcs to take in what string they need to filter for. Return a function from them that expects a message and, when it is called, filters based on the closed over value.
- Place all validation-based funcs into a List of funcs.
- Loop through each function with the same message and check the output.

Adjusting banned words needs work. Need to make sure things like ‘no’ and ’No’ are caught. They’re the same word just presented — orthographically, speaking — differently.
Also need to handle words that are modified by punctuation. If the banned word is ‘where’ and the message contains ‘where?’ then we should be able to create ‘*redacted*?’ from that.
r the commit message for your changes. Lines starting