Handle lowercase and punctuation in filtering and keyword functions + CreateValidationFunc.
The addition of the rules is so that functionality related to banned terms is handled properly.

Adjust KeywordInMessage to first search for words that match lowercase. This is so we’re careful not to say like ‘I’ and ‘I’m’ are the same (lowercase they’re not the same, nor are they the same after removal of punctuation) but we should definition catch things like ‘foxed’ and ‘foxed!’.

SanitiseMessage is doing a lot more than it did before.
The new changes: we have two checks.
First check is if the word and the banned word, when stripped of casing are the same. If yes, all good.
Second part is a more involved process to make sure we check words without punctuation and that, if they’re the same, that we redact the term AND add in the punctuation after. This calls on other functions to help it along.

We’ve also a new function CreateValidationFuncs that is used in the extract conversation.

p.s: this commit is bigger than it ought to be :/